Small 'bug' fix in StrawAccumulator class to take the sqrt of a double instead of integer
git-svn-id: http://svn.cern.ch/guest/lhcb/Alignment/trunk@89298 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2a38128e357ef1d17ede460ea2cc152ef487ee52

+++ Alignment/Calibration/OTCalibration/src/StrawAccumulator.h
@@ -46,7 +46,7 @@ namespace OTCal
     // proper units. however, the weight have a rather arbitrary scale
     // here. therefore, we'll just return rms/sqrt(N), sort of.
     double rms = std::sqrt( ( m_sumrawtimeresidualsq - m_sumrawtimeresidual*m_sumrawtimeresidual/m_sumw)/m_sumw ) ;
-    return rms / std::sqrt(m_nevents) ;
+    return rms / std::sqrt( (double) m_nevents) ;
   }
